+++ src/plugins/data_source/server/lib/error.test.ts
@@ -4,37 +4,37 @@
  */
 
 import { SavedObjectsErrorHelpers } from '../../../../../src/core/server';
-import { DataSourceError } from './error';
+import { createDataSourceError } from './error';
 
-describe('DataSourceError', () => {
-  it('create from savedObject bad request error should be 400 error', () => {
+describe('CreateDataSourceError', () => {
+  it('create from savedObject bad request error should generate 400 error', () => {
     const error = SavedObjectsErrorHelpers.createBadRequestError('test reason message');
-    expect(new DataSourceError('test prefix: ', error)).toMatchObject({
+    expect(createDataSourceError(error)).toMatchObject({
       statusCode: 400,
-      message: 'test prefix: test reason message: Bad Request',
+      message: 'Data Source Error: test reason message: Bad Request',
     });
   });
 
-  it('create from savedObject not found error should be 400 error', () => {
-    const error = SavedObjectsErrorHelpers.decorateNotAuthorizedError(new Error());
-    expect(new DataSourceError('test prefix: ', error)).toHaveProperty('statusCode', 400);
+  it('create from savedObject not found error should have statusCode 404', () => {
+    const error = SavedObjectsErrorHelpers.createGenericNotFoundError('type', 'id');
+    expect(createDataSourceError(error)).toHaveProperty('statusCode', 404);
   });
 
-  it('create from savedObject service unavailable error should be a 500 error', () => {
+  it('create from savedObject service unavailable error should have statusCode 503', () => {
     const error = SavedObjectsErrorHelpers.decorateOpenSearchUnavailableError(
       new Error('test reason message')
     );
-    expect(new DataSourceError('test prefix: ', error)).toMatchObject({
-      statusCode: 500,
-      message: 'test prefix: test reason message',
+    expect(createDataSourceError(error)).toMatchObject({
+      statusCode: 503,
+      message: 'Data Source Error: test reason message',
     });
   });
 
   it('create from non savedObject error should always be a 400 error', () => {
     const error = new Error('test reason message');
-    expect(new DataSourceError('test prefix: ', error)).toMatchObject({
+    expect(createDataSourceError(error)).toMatchObject({
       statusCode: 400,
-      message: 'test prefix: test reason message',
+      message: 'Data Source Error: test reason message',
     });
   });
 });